requirements.txt
google-cloud-aiplatform
google-cloud-firestore
google-cloud-storage
flask
pydantic
python-dotenv
pypdf2
vertexai